cleanup, refactoring
